Prereqs, security, formatting
Bot now makes an additional API call to be able to store and list prerequisites for queried classes. In addition to some small formatting changes, the config file now includes the auth for the concordia API that was previously public (oops). It is a new key value pair and the old one has been deleted.