Large reorganization
All functions now reside in the functions.py folder for better organization. Only the initial client event, main event, and task loops are in the newly renamed main.py.

Queried courses without a given semester (so len(user_message) = 2) now tells the user the course title as well as tells them which semester(s) the class is (are) offered in if they request a semester which it is not offered in. Much nicer than just being told it is not offered in that semester.